Renamed gui modules
