--- src/dat.js
@@ -1,4 +1,0 @@
-import * as dat from 'dat.gui';
-
-let gui = new dat.GUI();
-export { gui as datGUI }
