+++ src/objectManager.js
@@ -1,17 +1,12 @@
-
-import { loadnExecute } from './partLoader';
 import * as THREE from 'three';
 import OBJLoader2 from './extern/OBJLoader2'
 import seedrandom from 'seedrandom'
-import Snackbar from 'node-snackbar';
 
-let console= {
-    log:function (params) {
-        Snackbar.show({pos: 'bottom-right',text:params}); //Set the position
 
-    }
-    
-}
+
+import { loadnExecute } from './partLoader';
+import { OnScreen } from './gui';
+
 let THREEEX = Object.assign({}, THREE, { OBJLoader2, seedrandom });
 
 let maping = null;
@@ -80,7 +75,7 @@ export function loadUniverseAt(position, far, scene, setNeedToDisplay) {
                         scene.add(anchor);
                         item.object = anchor;
                         setNeedToDisplay();
-                        console.log(`Loaded  ${item.url}`);
+                        OnScreen.log(`Loaded  ${item.url}`);
 
                     })
 
@@ -107,7 +102,7 @@ export function unLoadUniverseAt(position, far, scene, setNeedToDisplay) {
         let distance = vectposition.distanceTo(position);
         if (distance - item.radius > far + 100 && item.object) {
 
-            console.log(`Unloading  ${item.url}`);
+            OnScreen.log(`Unloading  ${item.url}`);
             if (item.object.parent)
                 item.object.parent.remove(item.object);
             if (item.disposer) {
@@ -141,7 +136,7 @@ export function unLoadUniverseAt(position, far, scene, setNeedToDisplay) {
                 let box = new THREE.Box3().setFromObject(item.object);
                 let sphere = new THREE.Sphere();
                 box.getBoundingSphere(sphere);                
-                console.log(`${item.url} Eestimated radius: ${sphere.radius}`);
+                OnScreen.log(`${item.url} Eestimated radius: ${sphere.radius}`);
                 item.radius = sphere.radius;
